fix(swipe): use Long snackbar duration and emit UndoCompleted on undo
SnackbarDuration.Short (4s) was shorter than the 5s undo timeout,
creating a race condition. Switched to Long so the snackbar outlives
the undo window. Also emit UndoCompleted in undoDelete/undoArchive
so the snackbar dismisses immediately when the user taps Undo.

+++ composeApp/src/commonMain/kotlin/com/ricardocosteira/habitlock/presentation/ui/today/TodayScreen.kt
@@ -116,7 +116,7 @@ fun TodayScreen(
                     val result = snackbarHostState.showSnackbar(
                         message = getString(Res.string.swipe_habit_archived),
                         actionLabel = getString(Res.string.swipe_undo),
-                        duration = SnackbarDuration.Short
+                        duration = SnackbarDuration.Long
                     )
                     if (result == SnackbarResult.ActionPerformed) {
                         viewModel.undoArchive()
@@ -127,7 +127,7 @@ fun TodayScreen(
                     val result = snackbarHostState.showSnackbar(
                         message = getString(Res.string.swipe_habit_deleted),
                         actionLabel = getString(Res.string.swipe_undo),
-                        duration = SnackbarDuration.Short
+                        duration = SnackbarDuration.Long
                     )
                     if (result == SnackbarResult.ActionPerformed) {
                         viewModel.undoDelete()
